decrease memory usage by storing chunk colours as 8bit (at the cost of slower mesh generation)

+++ Sources/Voxelotl/Chunk.swift
@@ -123,7 +123,7 @@ public struct Chunk: Hashable {
 
 public enum BlockType: Hashable {
   case air
-  case solid(_ color: Color<Float>)
+  case solid(_ color: Color<UInt8>)
 }
 
 public struct Block: Hashable {
